Corrections on some chapters
Fix proof of the formal version of the Berry paradox.

Add clarification to the relation between primitive recursion and recursion on values strictly
smaller than the input.

Remove wrong statement about totality of f to have a r.e. pre-image from a r.e. set.

Fix proof that recursive A contains non-recursive r.e. B.

Add clarification on the role of h in the s-m-n property.